+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDBakedReflectionSystem.cs
@@ -131,26 +131,27 @@ namespace UnityEditor.Rendering.HighDefinition
             HashUtilities.AppendHash(ref skySettingsHash, ref allProbeDependencyHash);
 
             var bakedProbes = HDProbeSystem.bakedProbes;
+            var bakedProbeCount = HDProbeSystem.bakedProbeCount;
 
             // == 2. ==
-            var states = stackalloc HDProbeBakingState[bakedProbes.Count];
+            var states = stackalloc HDProbeBakingState[bakedProbeCount];
             ComputeProbeInstanceID(bakedProbes, states);
             ComputeProbeSettingsHashes(bakedProbes, states);
             // TODO: Handle bounce dependency here
-            ComputeProbeBakingHashes(bakedProbes.Count, allProbeDependencyHash, states);
+            ComputeProbeBakingHashes(bakedProbeCount, allProbeDependencyHash, states);
 
             CoreUnsafeUtils.QuickSort<HDProbeBakingState, Hash128, HDProbeBakingState.ProbeBakingHash>(
-                bakedProbes.Count, states
+                bakedProbeCount, states
             );
 
             int operationCount = 0, addCount = 0, remCount = 0;
-            var maxProbeCount = Mathf.Max(bakedProbes.Count, m_HDProbeBakedStates.Length);
+            var maxProbeCount = Mathf.Max(bakedProbeCount, m_HDProbeBakedStates.Length);
             var addIndices = stackalloc int[maxProbeCount];
             var remIndices = stackalloc int[maxProbeCount];
 
             if (m_HDProbeBakedStates.Length == 0)
             {
-                for (int i = 0; i < bakedProbes.Count; ++i)
+                for (int i = 0; i < bakedProbeCount; ++i)
                     addIndices[addCount++] = i;
                 operationCount = addCount;
             }
@@ -165,7 +166,7 @@ namespace UnityEditor.Rendering.HighDefinition
                             HDProbeBakingState, HDProbeBakingState.ProbeBakingHash
                        > (
                        m_HDProbeBakedStates.Length, oldBakedStates, // old hashes
-                       bakedProbes.Count, states,                   // new hashes
+                       bakedProbeCount, states,                   // new hashes
                        addIndices, remIndices,
                        out addCount, out remCount
                     );
@@ -202,7 +203,7 @@ namespace UnityEditor.Rendering.HighDefinition
                     // Get from cache or render the probe
                     if (!File.Exists(cacheFile))
                         RenderAndWriteToFile(probe, cacheFile, cubeRT, planarRT);
-                    
+
                     planarRT.Release();
                 }
                 cubeRT.Release();
@@ -232,7 +233,7 @@ namespace UnityEditor.Rendering.HighDefinition
                 for (int j = 0; j < 2; ++j)
                 {
                     AssetDatabase.StartAssetEditing();
-                    for (int i = 0; i < bakedProbes.Count; ++i)
+                    for (int i = 0; i < bakedProbeCount; ++i)
                     {
                         var index = addIndices[i];
                         var instanceId = states[index].instanceID;
@@ -312,7 +313,7 @@ namespace UnityEditor.Rendering.HighDefinition
             handle.SetIsDone(true);
         }
 
-        public static bool BakeProbes(IList<HDProbe> bakedProbes)
+        public static bool BakeProbes(IEnumerable<HDProbe> bakedProbes)
         {
             if (!(RenderPipelineManager.currentPipeline is HDRenderPipeline hdPipeline))
             {
@@ -326,9 +327,8 @@ namespace UnityEditor.Rendering.HighDefinition
             var cubeRT = HDRenderUtilities.CreateReflectionProbeRenderTarget(cubemapSize);
 
             // Render and write the result to disk
-            for (int i = 0; i < bakedProbes.Count; ++i)
+            foreach (var probe in bakedProbes)
             {
-                var probe = bakedProbes[i];
                 var bakedTexturePath = HDBakingUtilities.GetBakedTextureFilePath(probe);
                 var planarRT = HDRenderUtilities.CreatePlanarProbeRenderTarget((int)probe.resolution);
                 RenderAndWriteToFile(probe, bakedTexturePath, cubeRT, planarRT);
@@ -342,9 +342,8 @@ namespace UnityEditor.Rendering.HighDefinition
             for (int j = 0; j < 2; ++j)
             {
                 AssetDatabase.StartAssetEditing();
-                for (int i = 0; i < bakedProbes.Count; ++i)
+                foreach (var probe in bakedProbes)
                 {
-                    var probe = bakedProbes[i];
                     var bakedTexturePath = HDBakingUtilities.GetBakedTextureFilePath(probe);
                     AssetDatabase.ImportAsset(bakedTexturePath);
                     ImportAssetAt(probe, bakedTexturePath);
@@ -353,9 +352,8 @@ namespace UnityEditor.Rendering.HighDefinition
             }
 
             AssetDatabase.StartAssetEditing();
-            for (int i = 0; i < bakedProbes.Count; ++i)
+            foreach (var probe in bakedProbes)
             {
-                var probe = bakedProbes[i];
                 var bakedTexturePath = HDBakingUtilities.GetBakedTextureFilePath(probe);
 
                 // Get or create the baked texture asset for the probe
@@ -379,9 +377,8 @@ namespace UnityEditor.Rendering.HighDefinition
             // updateCount is a transient data, so don't execute this code before the asset reload.
             {
                 UnityEngine.Random.InitState((int)(1000 * hdPipeline.GetTime()));
-                for (int i = 0; i < bakedProbes.Count; ++i)
+                foreach (var probe in bakedProbes)
                 {
-                    var probe = bakedProbes[i];
                     var c = UnityEngine.Random.Range(2, 10);
                     while (probe.texture.updateCount < c) probe.texture.IncrementUpdateCount();
                 }
@@ -663,23 +660,28 @@ namespace UnityEditor.Rendering.HighDefinition
             return Path.Combine(hashFolder, string.Format("HDProbe-{0}.exr", hash));
         }
 
-        static void ComputeProbeInstanceID(IList<HDProbe> probes, HDProbeBakingState* states)
+        static void ComputeProbeInstanceID(IEnumerable<HDProbe> probes, HDProbeBakingState* states)
         {
-            for (int i = 0; i < probes.Count; ++i)
-                states[i].instanceID = probes[i].GetInstanceID();
+            var i = 0;
+            foreach (var probe in probes)
+            {
+                states[i].instanceID = probe.GetInstanceID();
+                ++i;
+            }
         }
 
-        static void ComputeProbeSettingsHashes(IList<HDProbe> probes, HDProbeBakingState* states)
+        static void ComputeProbeSettingsHashes(IEnumerable<HDProbe> probes, HDProbeBakingState* states)
         {
-            for (int i = 0; i < probes.Count; ++i)
+            var i = 0;
+            foreach (var probe in probes)
             {
-                var probe = probes[i];
                 var positionSettings = ProbeCapturePositionSettings.ComputeFrom(probe, null);
                 var positionSettingsHash = positionSettings.ComputeHash();
                 // TODO: make ProbeSettings and unmanaged type so its hash can be the hash of its memory
                 var probeSettingsHash = probe.settings.ComputeHash();
                 HashUtilities.AppendHash(ref positionSettingsHash, ref probeSettingsHash);
                 states[i].probeSettingsHash = probeSettingsHash;
+                ++i;
             }
         }