+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Shadowmask and realtime reflection probe property are hide in Quality settings
 - Made the StaticLightingSky class public so that users can change it by script for baking purpose.
 - Changed default exposure compensation to 0.
+- Improved performance of reflection probe management when using a lot of probes.
 
 ## [8.1.0] - 2020-04-21